The project is divided into modules
